refacotr: Break up components
