+++ apps/editor.planx.uk/src/pages/FlowEditor/components/forms/AddComponentModal/index.tsx
@@ -1,186 +1,21 @@
-import Box from "@mui/material/Box";
 import Popover from "@mui/material/Popover";
-import Tooltip from "@mui/material/Tooltip";
-import Typography from "@mui/material/Typography";
-import { ICONS } from "@planx/components/shared/icons";
 import { useNavigate, useParams } from "@tanstack/react-router";
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { useCallback } from "react";
 import type { NodeSearchParams } from "routes/_authenticated/app/$team/$flow/_flowEditor/nodes/route";
-import { FONT_WEIGHT_SEMI_BOLD } from "theme";
-import { AiChip } from "ui/editor/AiChip";
-import { SearchBox } from "ui/shared/SearchBox/SearchBox";
 import { getNodeRoute } from "utils/routeUtils/utils";
 
-import {
-  ALL_CATEGORIES,
-  ALL_ITEMS,
-  type Category,
-  type ComponentItem,
-} from "./ComponentsTab/componentData";
+import { ComponentsTab } from "./ComponentsTab";
 
 const POPOVER_WIDTH = 300;
 
-interface ComponentRowProps {
-  item: ComponentItem;
-  onClick: () => void;
-  scrollContainerRef: React.RefObject<HTMLDivElement | null>;
-}
-
-const ComponentRow: React.FC<ComponentRowProps> = ({
-  item,
-  onClick,
-  scrollContainerRef,
-}) => {
-  const Icon = ICONS[item.type];
-  const [tooltipOpen, setTooltipOpen] = useState(false);
-
-  useEffect(() => {
-    const el = scrollContainerRef.current;
-    if (!el) return;
-    const onScroll = () => setTooltipOpen(false);
-    el.addEventListener("scroll", onScroll, { passive: true });
-    return () => el.removeEventListener("scroll", onScroll);
-  }, [scrollContainerRef]);
-
-  return (
-    <Tooltip
-      title={item.description}
-      placement="right"
-      arrow
-      open={tooltipOpen}
-      onOpen={() => setTooltipOpen(true)}
-      onClose={() => setTooltipOpen(false)}
-      slotProps={{ tooltip: { sx: { maxWidth: 240 } } }}
-    >
-      <Box
-        onClick={onClick}
-        data-component-type={item.type}
-        sx={{
-          display: "flex",
-          alignItems: "center",
-          gap: 1,
-          px: 1.5,
-          py: 0.875,
-          cursor: "pointer",
-          "&:hover": {
-            backgroundColor: "action.hover",
-            "& .component-title": { fontWeight: FONT_WEIGHT_SEMI_BOLD },
-          },
-        }}
-      >
-        {Icon && (
-          <Box sx={{ flexShrink: 0, lineHeight: 0, color: "text.primary" }}>
-            <Icon sx={{ fontSize: 20 }} />
-          </Box>
-        )}
-        <Typography variant="body2" className="component-title">
-          {item.title}
-        </Typography>
-        {item.hasAiVariant && <AiChip sx={{ ml: "auto" }} />}
-      </Box>
-    </Tooltip>
-  );
-};
-
-interface AddComponentModalContentProps {
-  onSelect: (slug: string) => void;
-}
-
-export const AddComponentModalContent: React.FC<
-  AddComponentModalContentProps
-> = ({ onSelect }) => {
-  const [searchedItems, setSearchedItems] = useState<ComponentItem[] | null>(
-    null,
-  );
-  const listRef = useRef<HTMLDivElement>(null);
-
-  const filteredCategories = useMemo<Category[]>(() => {
-    if (!searchedItems) return ALL_CATEGORIES;
-    const visibleSlugs = new Set(searchedItems.map((item) => item.slug));
-    return ALL_CATEGORIES.map((cat) => ({
-      ...cat,
-      items: cat.items.filter((item) => visibleSlugs.has(item.slug)),
-    })).filter((cat) => cat.items.length > 0);
-  }, [searchedItems]);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      (document.getElementById("search") as HTMLInputElement | null)?.focus();
-    }, 50);
-    return () => clearTimeout(timer);
-  }, []);
-
-  return (
-    <>
-      <Box
-        sx={{
-          px: 1.5,
-          py: 1.25,
-          borderBottom: 1,
-          borderColor: "divider",
-          backgroundColor: "background.paper",
-        }}
-      >
-        <SearchBox<ComponentItem>
-          records={ALL_ITEMS}
-          setRecords={setSearchedItems}
-          searchKey={["title", "description"]}
-          compact
-          hideLabel
-          fullWidth
-          placeholder="Search components"
-        />
-      </Box>
-      <Box ref={listRef} sx={{ overflowY: "auto", pb: 2 }}>
-        {filteredCategories.length === 0 ? (
-          <Typography color="textSecondary" variant="body2" sx={{ p: 2 }}>
-            No components match your search.
-          </Typography>
-        ) : (
-          filteredCategories.map((cat) => (
-            <Box key={cat.label}>
-              <Typography
-                variant="body3"
-                sx={{
-                  fontWeight: FONT_WEIGHT_SEMI_BOLD,
-                  display: "block",
-                  p: 1.5,
-                  pb: 0.5,
-                  color: "text.primary",
-                }}
-              >
-                {cat.label}
-              </Typography>
-              {cat.items.map((item) => (
-                <ComponentRow
-                  key={item.slug}
-                  item={item}
-                  onClick={() => onSelect(item.slug)}
-                  scrollContainerRef={listRef}
-                />
-              ))}
-            </Box>
-          ))
-        )}
-      </Box>
-    </>
-  );
-};
-
-interface AddComponentModalProps {
+interface Props {
   anchorEl: HTMLElement | null;
   parent?: string;
   before?: string;
   onClose: () => void;
 }
 
-const AddComponentModal: React.FC<AddComponentModalProps> = ({
+const AddComponentModal: React.FC<Props> = ({
   anchorEl,
   parent,
   before,
@@ -243,7 +78,7 @@ const AddComponentModal: React.FC<AddComponentModalProps> = ({
         },
       }}
     >
-      <AddComponentModalContent onSelect={handleSelect} />
+      <ComponentsTab onSelect={handleSelect} />
     </Popover>
   );
 };
